pycon-app/requirements.txt
Flask==2.3.3
web3==6.11.3
python-dotenv==1.0.0
Werkzeug==2.3.7

